requirements/base.txt
# Django
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary

# Celery
celery
redis
django-celery-results
django-celery-beat

# Channels
channels
channels-redis
daphne

# Reporting
python-docx
openpyxl
reportlab
Pillow

# Other
django-cors-headers
django-mptt
PyJWT
djangorestframework-recursive
django-redis
gunicorn
